+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
+- `sandstone/Dropdown' to focus properly the first option and the last option via page up and page down
 - `sandstone/QuickGuidePanels` to not lose focus when the last view is displayed
 
 ## [2.7.5] - 2023-08-04